Backup only a list of files and subdirectories
In certain hosting configurations i have a wordpress site in the document root of the domain name and i have in subdirectories other sites belonging to subdomains of the same domain name. So i need to backup only the index.php file and the wp-* directories and files of the root directory. Furthermore, i usually have other directories storing other data in the same root directory of the main WP site...

This change in the SH script need an optional new parameter called file_list in the profile file for the wp site. Something like this:

file_list="wp-* favicon.ico index*"

if you comment this file then the default behaviour of the script is the same until now: backup all the directories and files inside the document root (wp_root parameter on the profile file).