Final optimization of CPU & time consuming, not compressing
I'm sorry, because i've needed to refactor my 2 previous commits. But finally it save almost 10 times the time needed to make a backup of about 7Gb of space. The solution was not only not merge the DB backup with the backup of the Home directory, but also NOT COMPRESS that Home directory copy. So instead of get a compressed TAR.GZ of the Home Directory we get a TAR file containing the Home Directory (or the files and subdirectories you specify in the profile).

Furthermore, when you specify not to compress i choosed to not to join both copies (DB + HOME DIR) and leave it as 2 different files per daily backup. At least for me has another advantage: i can get a copy of the database without decompress the daily backup.

My change introduced in this commit needs this in the profile (i will commit it in the next commit):

# Optional: if your files & directories to be backed up are too bigger in size, you can set this to false
# and then that backup will not be compressed and the the daily backup were composed by 2 files (SQL.GZ for DB and TAR for files&dirs)
compressed_tar_file=false