Optional parameter on profile: NO COMPRESSED TAR
This new OPTIONAL parameter is to make run my other previous commit, to get 2 FILES each day per WP site backed up (instead a unique compressed TAR.GZ), and the one containing the Home Directory is not compressed (so it is a TAR file, not TAR.GZ) and it save almost 10 times of processing time for a Home directory of about 7Gb with a lot of videos an photos. 

The resulting TAR file is maybe 20-35% bigger, but i prefer to get it faster than smaller. Space is not so a problem.